+++ lsm/utils.cpp
@@ -10,9 +10,9 @@ void die(const std::string& message) {
 };
 
 // Given a number of microseconds, return a string with the minutes and seconds to 2 significant digits.
-std::string formatMicroseconds(long long microseconds) {
-    long long totalSeconds = microseconds / 1000000;
-    long long minutes = totalSeconds / 60;
+std::string formatMicroseconds(size_t microseconds) {
+    size_t totalSeconds = microseconds / 1000000;
+    size_t minutes = totalSeconds / 60;
     microseconds %= 1000000;
     double seconds = static_cast<double>(totalSeconds % 60) + static_cast<double>(microseconds) / 1000000;
