+++ lsm/utils.hpp
@@ -4,7 +4,7 @@
 
 // Declarations of error and string handling functions
 void die(const std::string& message);
-std::string formatMicroseconds(long long microseconds);
+std::string formatMicroseconds(size_t microseconds);
 std::string addCommas(std::string s);
 
 #endif /* UTILS_HPP */
